Correct project structure (remove overengineering)
There is no point in having env_specific files.
There is no point of having dev/ and prod/ dirs as it is in mobile.
Why? In backend everything what is deployable is prod. There is no
distinction. So in this project dev is literally what serves the dev
locally, during the development on his machine.